backend/requirements.txt
aioredis==1.0.0
amqp==2.3.2
asgiref==2.2.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
bcrypt==3.1.6
billiard==3.5.0.4
celery==4.1.1
certifi==2018.1.18
cffi==1.12.2
channels==2.0.2
channels-redis==2.1.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==2.1.0
dateparser==0.7.0
Django==2.0.3
django-db-geventpool==3.0.1
django-rest-framework==0.1.0
django-storages==1.7.1
django-table-select-widget==1.1.0
django-tsvector-field==0.9.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
envparse==0.2.0
facebook-sdk==2.0.0
feedparser==5.2.1
future==0.16.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
kombu==4.2.1
msgpack==0.5.6
oauthlib==2.0.7
paramiko==2.4.2
psycogreen==1.0.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyaes==1.6.1
pyasn1==0.4.2
pycountry==18.5.26
pycparser==2.19
PyJWT==1.6.4
PyNaCl==1.3.0
PySocks==1.6.8
python-dateutil==2.7.2
python-gssapi==0.6.4
python-http-client==3.1.0
python-twitter==3.4.1
pytz==2018.3
redis==2.10.6
regex==2018.2.21
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
sendgrid==5.6.0
sentry-sdk==0.7.7
six==1.11.0
stripe==2.23.0
Telethon==1.6.2
tweepy==3.6.0
Twisted==17.9.0
txaio==2.9.0
tzlocal==1.5.1
urllib3==1.22
vine==1.1.4
zope.interface==4.4.3
git+https://github.com/willardmr/DjangoTableSelectMultipleWidget.git
django-celery-beat
channels_rabbitmq
lxml